adds grammar to set starting amount of sats per L2 node (#49)
This adds the ability to set starting amount of sats per L2 node within a doppler file. The example file to use is starting_amounts.doppler and the default amount used if not set is 10000000 sats. Additionally this PR pulls in an updated version of the parser library so coreln flags are correct added to config files (this allows us to finally use the dev-fast-gossip flag).