adds ability to command LND nodes using REST interface (#53)
This feature lays the ground work for us to be able to command nodes not running the in the doppler docker containers. Next step will be figuring out a good method to tell doppler to use different options than what are in the doppler containers to run commands against. Once this is done, we will be able to simulate activity against a remote cluster of nodes. This re-work was done in a way that it doesn't break any current functionality and LND nodes can still be commanded via lncli. The flag to use the rest option is --rest which needs to be run at the type doppler is kicked off.